+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/chat/room/club/dto/response/ClubChatRoomPreviewResponseDto.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.chat.clubChatRoom.dto.response;
+package com.mobble.mobbleserver.infrastructure.web.chat.room.club.dto.response;
 
 import com.mobble.mobbleserver.domain.chat.message.ChatMessage;
 import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
