+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/chat/room/direct/dto/request/DirectChatRoomCreateRequestDto.java
@@ -1,0 +1,4 @@
+package com.mobble.mobbleserver.infrastructure.web.chat.room.direct.dto.request;
+
+public record DirectChatRoomCreateRequestDto(Long receiverId) {
+}
