--- src/main/java/com/mobble/mobbleserver/refactor/chat/directChatRoom/dto/request/DirectChatRoomCreateRequestDto.java
@@ -1,4 +1,0 @@
-package com.mobble.mobbleserver.refactor.chat.directChatRoom.dto.request;
-
-public record DirectChatRoomCreateRequestDto(Long receiverId) {
-}
